Refactor callback.go to different go files
to prevent the file name is the same as package name,
separating the function to three different go files.
fix #237
